requirements.txt
absl-py==0.14.1
addict==2.4.0
aiohttp==3.7.4.post0
apex==0.1
async-timeout==3.0.1
attrs==21.2.0
av==8.0.2
backcall==0.2.0
Brotli==1.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
decord==0.6.0
easydict==1.9
einops==0.3.2
entrypoints==0.4
filelock==3.3.0
fonttools==4.38.0
fsspec==2021.10.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
horovod==0.28.1
huggingface-hub==0.16.4
idna==3.2
importlib-metadata==4.8.1
ipykernel==6.16.2
ipython==7.34.0
jedi==0.19.1
joblib==1.1.0
jsonlines==3.1.0
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.5
lmdb==0.97
lpips==0.1.4
lz4==2.1.9
Markdown==3.3.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
mmcv-full==1.3.10
multidict==5.2.0
mutagen==1.47.0
nest-asyncio==1.6.0
nltk==3.8.1
NRCLex==3.0.0
numpy==1.21.2
nvidia-dali-cuda100==1.3.0
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.43
protobuf==3.18.1
psutil==5.8.0
ptyprocess==0.7.0
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.20.0
pyDeprecate==0.3.1
Pygments==2.17.2
PyGObject==3.26.1
pyparsing==2.4.7
pyspark==2.4.7
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.9
PyYAML==5.4.1
pyzmq==25.1.2
regex==2021.10.8
requests==2.31.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
safetensors==0.4.2
scikit-learn==1.0.2
scipy==1.7.1
sentencepiece==0.1.96
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.0
textblob==0.17.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.12.0+cu113
torchaudio==0.12.0
torchmetrics==0.5.1
torchvision==0.13.0
tornado==6.2
tqdm==4.62.3
traitlets==5.9.0
transformers==4.30.2
typing==3.7.4.3
typing_extensions==4.7.1
ujson==4.2.0
unattended-upgrades==0.1
urllib3==2.0.7
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==2.0.2
yapf==0.31.0
yarl==1.7.0
youtube-dl==2021.12.17
yt-dlp==2023.11.16
zipp==3.6.0
